added horizontal movements
